Split support code out of large Swift files
Extract focused support code from ContentView.swift and cmuxApp.swift into smaller Swift files without intended behavior changes.
