+++ Sources/Sidebar/SidebarState.swift
@@ -1,0 +1,31 @@
+import Combine
+import CoreGraphics
+
+final class SidebarState: ObservableObject {
+    @Published var isVisible: Bool
+    @Published var persistedWidth: CGFloat
+
+    init(isVisible: Bool = true, persistedWidth: CGFloat = CGFloat(SessionPersistencePolicy.defaultSidebarWidth)) {
+        self.isVisible = isVisible
+        let sanitized = SessionPersistencePolicy.sanitizedSidebarWidth(Double(persistedWidth))
+        self.persistedWidth = CGFloat(sanitized)
+    }
+
+    func toggle() {
+        isVisible.toggle()
+    }
+}
+
+enum SidebarResizeInteraction {
+    // Keep a generous drag target inside the sidebar itself, but make the
+    // terminal-side overlap very small so column-0 text selection still wins.
+    static let sidebarSideHitWidth: CGFloat = 6
+    // 4 pt matches the 4 pt padding used in GhosttySurfaceScrollView drop zone overlays
+    // (dropZoneOverlayFrame). This prevents column-0 text near the leading edge from
+    // accidentally triggering the sidebar resize when interacting with leftmost content.
+    static let contentSideHitWidth: CGFloat = 4
+
+    static var totalHitWidth: CGFloat {
+        sidebarSideHitWidth + contentSideHitWidth
+    }
+}
